move HTTP related methods from utils to http-utils.
