+++ src/utils.js
@@ -14,13 +14,10 @@ import fs from './fs-promise';
 import Async from 'async';
 import Path from 'path';
 import AsyncLock from 'async-lock';
-import HttpProxyAgent from 'http-proxy-agent';
-import HttpsProxyAgent from 'https-proxy-agent';
 
 import { DefaultValues } from './constants';
 import UploadError from './upload-error';
 import ErrorCodes from './error-codes';
-import UploadFile from './upload-file';
 
 const lock = new AsyncLock();
 
@@ -430,78 +427,4 @@ export async function walkDirectory(directoryPath, maximumPaths = 5000, includeD
  */
 export async function getLock(lockId, callback) {
     return lock.acquire(lockId, callback);
-}
-
-/**
- * Builds proxy agent options based on upload options. Note that the method may return a falsy value, which
- * indicates that a proxy does not apply.
- * @param {DirectBinaryUploadOptions} directBinaryUploadOptions Options from which to retrieve information.
- * @returns {object} Options for either http-proxy-agent or https-proxy-agent.
- */
-export function getProxyAgentOptions(directBinaryUploadOptions) {
-    const proxy = directBinaryUploadOptions.getHttpProxy();
-    if (proxy) {
-        const proxyOptions = proxy.getUrl();
-        const user = proxy.getBasicAuthUser();
-        const password = proxy.getBasicAuthPassword();
-        if (user) {
-            proxyOptions.auth = `${user}:${password}`;
-        }
-        return proxyOptions;
-    }
-    return false;
-}
-
-/**
- * Converts options provided in a DirectBinaryUploadOptions instance to a format
- * suitable to pass to the httptransfer module.
- * @param {object} options General upload object options.
- * @param {DirectBinaryUploadOptions} directBinaryUploadOptions Options to convert.
- */
-export function getHttpTransferOptions(options, directBinaryUploadOptions) {
-    // the httptransfer module accepts a full fileUrl instead of a single
-    // url with individual file names. if needed, convert the format with a
-    // single url and individual file names to the fileUrl format.
-    const convertedFiles = directBinaryUploadOptions.getUploadFiles().map((uploadFile) => {
-        const uploadFileInstance = new UploadFile(options, directBinaryUploadOptions, uploadFile);
-        const transferOptions = uploadFileInstance.toJSON();
-        if (uploadFile.blob) {
-            // ensure blob is passed through to transfer options
-            transferOptions.blob = uploadFile.blob;
-        }
-        return transferOptions;
-    });
-
-    const transferOptions = {
-        uploadFiles: convertedFiles,
-        headers: directBinaryUploadOptions.getHeaders(),
-        concurrent: directBinaryUploadOptions.isConcurrent(),
-        maxConcurrent: directBinaryUploadOptions.getMaxConcurrent(),
-    };
-
-    const proxyOptions = getProxyAgentOptions(directBinaryUploadOptions);
-    if (proxyOptions) {
-        const { protocol = 'http:' } = proxyOptions;
-        transferOptions.requestOptions = {
-            agent: protocol === 'https:' ? new HttpsProxyAgent(proxyOptions) : new HttpProxyAgent(proxyOptions)
-        };
-    }
-
-    return transferOptions;
-}
-
-/**
- * Validates and retrieves basic authentication information from a set of options. Will throw an error
- * if only one of username or password is provided.
- * @param {object} options Options from which "username" and "password" properties will be retrieved.
- */
-export function getBasicAuth(options) {
-    const { username, password } = options;
-    if (username && !password) {
-        throw new UploadError('password is required for basic auth', ErrorCodes.INVALID_OPTIONS);
-    }
-    if (password && !username) {
-        throw new UploadError('username is required for basic auth', ErrorCodes.INVALID_OPTIONS);
-    }
-    return options;
 }
